restructure sensitivity plotting function into own module
