+++ aixcalibuha/sensitivity_analysis/__init__.py
@@ -2,8 +2,10 @@
 This package contains classes to perform
 sensitivity analysis with.
 """
-from .sensitivity_analyzer import SenAnalyzer, _del_duplicates, _rename_tuner_names
+
+from .sensitivity_analyzer import SenAnalyzer
 from .sobol import SobolAnalyzer
 from .morris import MorrisAnalyzer
 from .fast import FASTAnalyzer
 from .pawn import PAWNAnalyzer
+import aixcalibuha.sensitivity_analysis.plotting as plotting
