+++ aixcalibuha/sensitivity_analysis/sensitivity_analyzer.py
@@ -13,11 +13,8 @@ from ebcpy import data_types
 from ebcpy.simulationapi import SimulationAPI
 from aixcalibuha import CalibrationClass, data_types
 from aixcalibuha import utils
-import matplotlib.pyplot as plt
-from SALib.plotting.bar import plot as barplot
 import multiprocessing as mp
-import time
-import warnings
+from aixcalibuha.sensitivity_analysis.plotting import plot_single, plot_time_dependent
 
 
 def _load_single_file(_filepath, parquet_engine='pyarrow'):
@@ -34,26 +31,6 @@ def _load_files(_filepaths, parquet_engine='pyarrow'):
     for _filepath in _filepaths:
         results.append(_load_single_file(_filepath, parquet_engine=parquet_engine))
     return results
-
-
-def _del_duplicates(x):
-    """Helper function"""
-    return list(dict.fromkeys(x))
-
-
-def _get_suffix(modelica_var_name):
-    """Helper function"""
-    index_last_dot = modelica_var_name.rfind('.')
-    suffix = modelica_var_name[index_last_dot + 1:]
-    return suffix
-
-
-def _rename_tuner_names(result):
-    """Helper function"""
-    tuner_names = list(result.columns)
-    rename_tuner_names = {name: _get_suffix(name) for name in tuner_names}
-    result = result.rename(columns=rename_tuner_names, index=rename_tuner_names)
-    return result
 
 
 def _restruct_verbose(list_output_verbose):
@@ -748,9 +725,9 @@ class SenAnalyzer(abc.ABC):
             self._save(sen_time_dependent_df, time_dependent=True)
         if plot_result:
             if isinstance(sen_time_dependent_df, pd.DataFrame):
-                self.plot_time_dependent(sen_time_dependent_df)
+                plot_time_dependent(sen_time_dependent_df)
             else:
-                self.plot_time_dependent(sen_time_dependent_df[0])
+                plot_time_dependent(sen_time_dependent_df[0])
         return sen_time_dependent_df
 
     def _analyze_tstep_df(self, time_step, tsteps_sim_results, variables, samples, cal_class):
@@ -838,264 +815,15 @@ class SenAnalyzer(abc.ABC):
         """
         raise NotImplementedError
 
-    @staticmethod
-    def plot_single(result: pd.DataFrame, **kwargs):
-        """
-        Plot sensitivity results of first and total order analysis variables.
-        For each calibration class one figure is created, which shows for each goal an axis
-        with a barplot of the values of the analysis variables.
-
-        :param pd.DataFrame result:
-            A result from run
-        :keyword bool show_plot:
-            Default is True. If False, all created plots are not shown.
-        :keyword bool use_suffix:
-            Default is True: If True, the last part after the last point
-            of Modelica variables is used for the x ticks.
-        :keyword [str] cal_classes:
-            Default are all possible calibration classes. If a list of
-            names of calibration classes is given only plots for these
-            classes are created.
-        :keyword [str] goals:
-            Default are all possible goal names. If a list of specific
-            goal names is given only these will be plotted.
-        :keyword ([fig], [ax]) figs_axes:
-            Default None. Useful for using subfigures (see example for verbose sensitivity analysis).
-        :return:
-            Returns all created figures and axes in lists like [fig], [ax]
-        """
-        show_plot = kwargs.pop('show_plot', True)
-        # kwargs for the design
-        use_suffix = kwargs.pop('use_suffix', False)
-        figs_axes = kwargs.pop('figs_axes', None)
-
-        # get lists of the calibration classes and their goals in the result dataframe
-        cal_classes = kwargs.pop('cal_classes', None)
-        if cal_classes is None:
-            cal_classes = _del_duplicates(list(result.index.get_level_values(0)))
-        goals = kwargs.pop('goals', None)
-        if goals is None:
-            goals = _del_duplicates(list(result.index.get_level_values(1)))
-
-        # rename tuner_names in result to the suffix of their variable name
-        if use_suffix:
-            result = _rename_tuner_names(result)
-
-        # when the index is not sorted pandas throws a performance warning
-        result = result.sort_index()
-
-        # plotting with simple plot function of the SALib
-        figs = []
-        axes = []
-        for col, cal_class in enumerate(cal_classes):
-            if figs_axes is None:
-                fig, ax = plt.subplots(len(goals), sharex='all')
-            else:
-                fig = figs_axes[0][col]
-                ax = figs_axes[1][col]
-            fig.suptitle(cal_class)
-            figs.append(fig)
-            if not isinstance(ax, np.ndarray):
-                ax = [ax]
-            axes.append(ax)
-            for row, goal in enumerate(goals):
-                result_df = result.loc[cal_class, goal]
-                axes[col][row].grid(True, which='both', axis='y')
-                barplot(result_df.T, ax=axes[col][row])
-                axes[col][row].set_title(goal)
-                axes[col][row].legend()
-
-        if show_plot:
-            plt.show()
-
-        return figs, axes
-
-    @staticmethod
-    def plot_time_dependent(result: pd.DataFrame, **kwargs):
-        """
-        Plot time dependent sensitivity results without interactions from run_time_dependent().
-
-        For each goal one figure is created with one axes for each analysis variable.
-        In these plots the time dependent sensitivity of the parameters is plotted.
-        The confidence interval can also be plotted.
-
-        :param pd.DataFrame result:
-            A result from run_time_dependent without second order results.
-        :keyword [str] parameters:
-            Default all parameters. List of parameters to plot the sensitivity.
-        :keyword bool plot_conf:
-            Default True. If true, the confidence intervals for each parameter are plotted.
-        :keyword bool show_plot:
-            Default is True. If False, all created plots are not shown.
-        :keyword bool use_suffix:
-            Default is True: If True, the last part after the last point
-            of Modelica variables is used for the x ticks.
-        :keyword [str] goals:
-            Default are all possible goal names. If a list of specific
-            goal names is given only these will be plotted.
-        :keyword ([fig], [ax]) figs_axes:
-            Default None. Optional custom figures and axes (see example for verbose sensitivity analysis).
-        :return:
-            Returns all created figures and axes in lists like [fig], [ax]
-        """
-        plot_conf = kwargs.pop('plot_conf', True)
-        show_plot = kwargs.pop('show_plot', True)
-        # kwargs for the design
-        use_suffix = kwargs.pop('use_suffix', False)
-        figs_axes = kwargs.pop('figs_axes', None)
-
-        goals = kwargs.pop('goals', None)
-        if goals is None:
-            goals = _del_duplicates(list(result.index.get_level_values(0)))
-        all_analysis_variables = _del_duplicates(list(result.index.get_level_values(1)))
-        analysis_variables = kwargs.pop('analysis_variables', None)
-        if analysis_variables is None:
-            analysis_variables = [av for av in all_analysis_variables if '_conf' not in av]
-
-        # rename tuner_names in result to the suffix of their variable name
-        if use_suffix:
-            result = _rename_tuner_names(result)
-
-        # when the index is not sorted pandas throws a performance warning
-        result = result.sort_index()
-        parameters = kwargs.pop('parameters', result.columns.values)
-
-        figs = []
-        axes = []
-        for g_i, goal in enumerate(goals):
-            if figs_axes is None:
-                fig, ax = plt.subplots(len(analysis_variables), sharex='all')
-            else:
-                fig = figs_axes[0][g_i]
-                ax = figs_axes[1][g_i]
-            fig.suptitle(goal)
-            figs.append(fig)
-            if not isinstance(ax, np.ndarray):
-                ax = [ax]
-            axes.append(ax)
-            for av_i, av in enumerate(analysis_variables):
-                axes[g_i][av_i].plot(result.loc[goal, av][parameters])
-                axes[g_i][av_i].set_ylabel(av)
-                axes[g_i][av_i].legend(parameters)
-                if plot_conf and av + '_conf' in all_analysis_variables:
-                    for p in parameters:
-                        y = result.loc[goal, av][p]
-                        x = y.index.to_numpy()
-                        ci = result.loc[goal, av + '_conf'][p]
-                        large_values_indices = ci[ci > 1].index
-                        if list(large_values_indices):
-                            warnings.warn(
-                                f"Confidence interval for {goal}, {av}, {p} was at the "
-                                f"following times {list(large_values_indices)} lager than 1 "
-                                f"and is smoothed out in the plot.")
-                        for idx in large_values_indices:
-                            prev_idx = ci.index.get_loc(idx) - 1
-                            if prev_idx >= 0:
-                                ci.iloc[ci.index.get_loc(idx)] = ci.iloc[prev_idx]
-                            else:
-                                ci.iloc[ci.index.get_loc(idx)] = 1
-                        axes[g_i][av_i].fill_between(x, (y - ci), (y + ci), alpha=.1)
-            axes[g_i][-1].set_xlabel('time')
-        if show_plot:
-            plt.show()
-        return figs, axes
-
-    @staticmethod
-    def plot_parameter_verbose(parameter, single_result, second_order_result=None, **kwargs):
-        """
-        Plot all time dependent sensitivity measure for one parameter.
-        For each goal an axes is created within one figure.
-
-        If second_order_results form SobolAnalyzer.run_time_dependent are given
-        the S2 results of the interaction with each other parameter are added on top
-        of each other and the first order result.
-
-        :param str parameter:
-            Parameter to plot all sensitivity results for. If use_suffix=True, then
-            the name must also be only the suffix.
-        :param pd.DataFrame single_result:
-            First and total order result form run_time_dependent.
-        :param pd.DataFrame second_order_result:
-            Default None. Second order result of SobolAnalyzer.run_time_dependent.
-        :keyword bool show_plot:
-            Default is True. If False, all created plots are not shown.
-        :keyword bool use_suffix:
-            Default is True: If True, the last part after the last point
-            of Modelica variables is used for the x ticks.
-        :keyword [str] goals:
-            Default are all possible goal names. If a list of specific
-            goal names is given only these will be plotted.
-        :keyword (fig, [ax]) fig_axes:
-            Default None. Optional custom figures and axes (see example for verbose sensitivity analysis).
-        :return:
-            Returns all created figures and axes in lists like [fig], [ax]
-        """
-        show_plot = kwargs.pop('show_plot', True)
-        # kwargs for the design
-        use_suffix = kwargs.pop('use_suffix', False)
-        fig_axes = kwargs.pop('fig_axes', None)
-
-        goals = kwargs.pop('goals', None)
-        if goals is None:
-            goals = _del_duplicates(list(single_result.index.get_level_values(0)))
-        all_analysis_variables = _del_duplicates(list(single_result.index.get_level_values(1)))
-        analysis_variables = [av for av in all_analysis_variables if '_conf' not in av]
-
-        # rename tuner_names in result to the suffix of their variable name
-        if use_suffix:
-            single_result = _rename_tuner_names(single_result)
-            # when the index is not sorted pandas throws a performance warning
-            single_result = single_result.sort_index()
-            if second_order_result is not None:
-                second_order_result = _rename_tuner_names(second_order_result)
-                second_order_result = second_order_result.sort_index()
-
-        if fig_axes is None:
-            fig, ax = plt.subplots(len(goals), sharex='all')
-        else:
-            fig = fig_axes[0]
-            ax = fig_axes[1]
-        fig.suptitle(parameter)
-        if not isinstance(ax, np.ndarray):
-            ax = [ax]
-        for g_i, goal in enumerate(goals):
-            if second_order_result is not None:
-                result_2_goal = second_order_result.loc[goal, 'S2', parameter]
-                mean = result_2_goal.mean().drop([parameter])
-                mean.sort_values(ascending=False, inplace=True)
-                sorted_interactions = list(mean.index)
-                time_ar = _del_duplicates(list(result_2_goal.index.get_level_values(0)))
-                value = single_result.loc[goal, 'S1'][parameter].to_numpy()
-                ax[g_i].plot(single_result.loc[goal, 'S1'][parameter], label='S1')
-                ax[g_i].fill_between(time_ar, np.zeros_like(value), value, alpha=0.1)
-                for para in sorted_interactions:
-                    value_2 = value + result_2_goal[para].to_numpy()
-                    ax[g_i].plot(time_ar, value_2, label='S2 ' + para)
-                    ax[g_i].fill_between(time_ar, value, value_2, alpha=0.1)
-                    value = value_2
-                ax[g_i].plot(single_result.loc[goal, 'ST'][parameter], label='ST')
-                legend = ['S1']
-                legend.extend(analysis_variables)
-                legend.append('ST')
-                ax[g_i].set_title(goal)
-                ax[g_i].legend()
-            else:
-                for av_i, av in enumerate(analysis_variables):
-                    ax[g_i].plot(single_result.loc[goal, av][parameter])
-                ax[g_i].legend(analysis_variables)
-        if show_plot:
-            plt.show()
-        return fig, ax
-
     def plot(self, result):
         """
         Plot the results of the sensitivity analysis method from run().
 
         :param pd.DataFrame result:
             Dataframe of the results like from the run() function.
-        :return tuple of matplotlib objects (fig, ax)
+        :return tuple of matplotlib objects (fig, ax):
         """
-        SenAnalyzer.plot_single(result=result)
+        plot_single(result=result)
 
     @staticmethod
     def load_from_csv(path):
